more event checks (#2670)
