+++ packages/common/src/channel/content-script.ts
@@ -4,6 +4,7 @@
 //
 
 import { isMobile } from "@coral-xyz/common-public";
+import { isValidEventOrigin } from "index";
 
 import { BrowserRuntimeCommon, BrowserRuntimeExtension } from "../browser";
 import { POST_MESSAGE_ORIGIN } from "../constants";
@@ -15,6 +16,9 @@ export class ChannelContentScript {
   // Forwards all messages from the client to the background script.
   public static proxy(reqChannel: string, respChannel: string) {
     window.addEventListener("message", (event) => {
+      if (!isValidEventOrigin(event)) {
+        return;
+      }
       if (event.data.type !== reqChannel) return;
       // @ts-ignore
       BrowserRuntimeCommon.sendMessageToAnywhere(
